refactor: domain 내부 패키지 분리
