+++ src/main/java/kitchenpos/application/OrderResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.application;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.Order;
+import kitchenpos.domain.order.Order;
 
 public class OrderResponse {
     private Long id;
