+++ src/main/java/kitchenpos/application/TableGroupResponse.java
@@ -3,7 +3,7 @@ package kitchenpos.application;
 import java.time.LocalDateTime;
 import java.util.List;
 import java.util.stream.Collectors;
-import kitchenpos.domain.TableGroup;
+import kitchenpos.domain.table.TableGroup;
 
 public class TableGroupResponse {
