+++ src/main/java/kitchenpos/domain/generic/Price.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.generic;
 
 import java.math.BigDecimal;
 import java.util.Objects;
