+++ src/main/java/kitchenpos/domain/menu/MenuGroup.java
@@ -1,8 +1,6 @@
-package kitchenpos.domain;
+package kitchenpos.domain.menu;
 
-import java.math.BigDecimal;
 import javax.persistence.Column;
-import javax.persistence.Embedded;
 import javax.persistence.Entity;
 import javax.persistence.GeneratedValue;
 import javax.persistence.GenerationType;
@@ -10,8 +8,8 @@ import javax.persistence.Id;
 import javax.persistence.Table;
 
 @Entity
-@Table(name = "product")
-public class Product {
+@Table(name = "menu_group")
+public class MenuGroup {
 
     @Id
     @GeneratedValue(strategy = GenerationType.IDENTITY)
@@ -20,20 +18,16 @@ public class Product {
     @Column(nullable = false)
     private String name;
 
-    @Embedded
-    private Price price;
-
-    protected Product() {
+    protected MenuGroup() {
     }
 
-    public Product(Long id, String name, BigDecimal price) {
+    public MenuGroup(Long id, String name) {
         this.id = id;
         this.name = name;
-        this.price = new Price(price);
     }
 
-    public Product(String name, BigDecimal price) {
-        this(null, name, price);
+    public MenuGroup(String name) {
+        this(null, name);
     }
 
     public Long getId() {
@@ -42,9 +36,5 @@ public class Product {
 
     public String getName() {
         return name;
-    }
-
-    public BigDecimal getPrice() {
-        return price.getValue();
     }
 }
