+++ src/main/java/kitchenpos/domain/table/OrderTableRepository.java
@@ -1,4 +1,4 @@
-package kitchenpos.domain;
+package kitchenpos.domain.table;
 
 import org.springframework.data.jpa.repository.JpaRepository;
