+++ src/test/java/kitchenpos/fixture/MenuBuilder.java
@@ -3,8 +3,8 @@ package kitchenpos.fixture;
 import java.math.BigDecimal;
 import java.util.ArrayList;
 import java.util.List;
-import kitchenpos.domain.Menu;
-import kitchenpos.domain.MenuProduct;
+import kitchenpos.domain.menu.Menu;
+import kitchenpos.domain.menu.MenuProduct;
 
 public class MenuBuilder {
