+++ src/test/java/kitchenpos/fixture/MenuGroupFactory.java
@@ -1,7 +1,7 @@
 package kitchenpos.fixture;
 
 import kitchenpos.application.MenuGroupRequest;
-import kitchenpos.domain.MenuGroup;
+import kitchenpos.domain.menu.MenuGroup;
 
 public class MenuGroupFactory {
