+++ src/test/java/kitchenpos/fixture/OrderTableFactory.java
@@ -1,6 +1,6 @@
 package kitchenpos.fixture;
 
-import kitchenpos.domain.OrderTable;
+import kitchenpos.domain.table.OrderTable;
 
 public class OrderTableFactory {
